+++ migration-library/src/main/kotlin/com/quadient/migration/api/dto/migrationmodel/builder/DocumentContentBuilderBase.kt
@@ -12,6 +12,7 @@ import com.quadient.migration.api.dto.migrationmodel.AttachmentRef
 import com.quadient.migration.api.dto.migrationmodel.StringValue
 import com.quadient.migration.api.dto.migrationmodel.Variable
 import com.quadient.migration.api.dto.migrationmodel.VariableRef
+import com.quadient.migration.api.dto.migrationmodel.builder.documentcontent.ColumnLayoutBuilder
 import com.quadient.migration.api.dto.migrationmodel.builder.documentcontent.RepeatedContentBuilder
 import com.quadient.migration.api.dto.migrationmodel.builder.documentcontent.SelectByLanguageBuilder
 import com.quadient.migration.shared.LiteralPath
@@ -242,4 +243,13 @@ interface DocumentContentBuilderBase<T> {
      */
     fun repeatedContent(variable: Variable, builder: RepeatedContentBuilder.() -> Unit): T =
         repeatedContent(VariableRefPath(variable.id), builder)
+
+    /**
+     * Defines a column layout modifier that affects sibling content within the current block.
+     * @param builder A builder function to configure the [ColumnLayoutBuilder].
+     * @return This builder instance for method chaining.
+     */
+    fun columnLayout(builder: ColumnLayoutBuilder.() -> Unit = {}): T = apply {
+        this.content.add(ColumnLayoutBuilder().apply(builder).build())
+    } as T
 }