+++ migration-library/src/main/kotlin/com/quadient/migration/persistence/migrationmodel/ColumnLayoutEntity.kt
@@ -1,0 +1,14 @@
+package com.quadient.migration.persistence.migrationmodel
+
+import com.quadient.migration.shared.ColumnApplyTo
+import com.quadient.migration.shared.ColumnBalancingType
+import com.quadient.migration.shared.Size
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ColumnLayoutEntity(
+    val numberOfColumns: Int,
+    val gutterWidth: Size? = null,
+    val balancingType: ColumnBalancingType? = null,
+    val applyTo: ColumnApplyTo? = null,
+) : DocumentContentEntity, TextContentEntity