+++ migration-library/src/main/kotlin/com/quadient/migration/shared/ColumnApplyTo.kt
@@ -1,0 +1,9 @@
+package com.quadient.migration.shared
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+enum class ColumnApplyTo {
+    WholeTemplate,
+    ThisBlockOnly
+}